requirements.txt
decorator>=4.0.11
imageio>=2.1.2
moviepy>=0.2.3.2
numpy>=1.14.0
Pillow>=5.0.0
tqdm>=4.11.2
Wand>=0.4.4
tkcolorpicker
